polymarket_trading_system/requirements.txt
# Core Dependencies
python-web3==6.15.1
py-clob-client==0.21.0
requests==2.31.0
websockets==12.0
asyncio-mqtt==0.16.1

# Database & ORM
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Configuration & Environment
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Monitoring & Logging
flask==3.0.0
prometheus-client==0.19.0
structlog==23.2.0
colorlog==6.8.0

# Data Analysis & Visualization
pandas==2.1.4
numpy==1.26.2
plotly==5.17.0
dash==2.16.1

# Security & Encryption
cryptography==42.0.0
keyring==24.3.0

# Networking & APIs
httpx==0.26.0
aiohttp==3.9.1
websocket-client==1.7.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-mock==3.12.0

# Utilities
click==8.1.7
python-dateutil==2.8.2
schedule==1.2.0
python-telegram-bot==20.7

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0